(refactored) moved to display component
